updated electron backend, cleaning up codebase and adding comments for next iteration team. Added cloudbased.ts to modularize backend logic for AWS instances. Co-authored-by: Jon Cruz <Jrcrz@users.noreply.github.com> Co-authored-by: Elena Atencio <elenaatencio@users.noreply.github.com> Co-authored-by: Iris Wong <wiris316@users.noreply.github.com> Co-authored-by: John Donato <jdonuto@users.noreply.github.com>
