+++ app/context/ApplicationContext.tsx
@@ -9,6 +9,7 @@ const { ipcRenderer } = window.require('electron');
  * @property  {Array} servicesData The microservices of that application
  * @method    fetchServicesNames
  * @method    connectToDB
+ * @method    getSavedMetrics
  */
 
 interface AppContextProps {
@@ -20,59 +21,32 @@ export const ApplicationContext = React.createContext<any>(null);
 const ApplicationContextProvider: React.FC<AppContextProps> = React.memo(props => {
   const children = props.children;
   const [servicesData, setServicesData] = useState([]);
-  const [serviceConnected, setServiceConnected] = useState<boolean>(false);
-  // v10 note: there is not function for setApp so app is never updated.
   const [app, setApp] = useState<string>('');
   const [savedMetrics, setSavedMetrics] = useState({});
   const [appIndex, setAppIndex] = useState<number>(0);
   const [intervalID, setIntervalID] = useState<NodeJS.Timeout | null>(null);
   
-  /**
- * @function tryParseJson - a function that will parse the JSON data into an object. If there is any error during parsing (a.k.a there is a circular inside the JSON data), console log the error
- * @param jsonString - JSON data that is received from backend
- * @return an object with all information from backend
- */
-  function tryParseJSON(jsonString: string) {
-    try {
-      const o = JSON.parse(jsonString);
-
-      if (o && typeof o === 'object') {
-        return o;
-      }
-    } catch (e) {
-      console.log({ error: e });
-    }
-    return false;
-  }
 
   /**
    * @function fetchServicesNames - a function that will take an application name and update the state of `serviceData` based on backend response
    * 1. Take in an application name
-   * 
    * 2. Send a `servicesRequest` to backend
-   * 
    * 3. Upon `servicesResponse`, parse the received JSON data and assign it to `servicesData`
-   * 
    * 4. Remove the listener for `servicesResponse`
    * @param application - application name
    */  
   // v10: Invoked by connectToDB, passing in app (card title)
   const fetchServicesNames = useCallback((application: string) => {
-    console.log('Hi, inside ApplicationConext - fetchServicesNames callback. Sending servicesRequest to ipcMain.');
-    console.log('app when fetch services name: ', application);
-    // setApp(application);
-
+    // console.log('Hi, inside ApplicationConext - fetchServicesNames callback. Sending servicesRequest to ipcMain.');
+    // console.log('app when fetch services name: ', application);
     ipcRenderer.send('servicesRequest');
-
     ipcRenderer.on('servicesResponse', (event: Electron.Event, data: any) => {
       let result: any;
-      // Parse JSON data into object
-      // if (tryParseJSON(data)) result = JSON.parse(data);
       result = JSON.parse(data);
-      console.log('result from ipcrenderer services response is: ', result);
-      console.log('Calling setServicesData passing in above result. Current servicesData is the following: ', servicesData);
+      // console.log('result from ipcrenderer services response is: ', result);
+      // console.log('Calling setServicesData passing in above result. Current servicesData is the following: ', servicesData);
       setServicesData(result);
-      console.log('Leaving fetchedServicesNames function.');
+      // console.log('Leaving fetchedServicesNames function.');
       ipcRenderer.removeAllListeners('servicesResponse');
     });
 
@@ -138,3 +112,22 @@ const ApplicationContextProvider: React.FC<AppContextProps> = React.memo(props =
 });
 
 export default ApplicationContextProvider;
+
+
+//   /**
+//  * @function tryParseJson - a function that will parse the JSON data into an object. If there is any error during parsing (a.k.a there is a circular inside the JSON data), console log the error
+//  * @param jsonString - JSON data that is received from backend
+//  * @return an object with all information from backend
+//  */
+//   function tryParseJSON(jsonString: string) {
+//     try {
+//       const o = JSON.parse(jsonString);
+
+//       if (o && typeof o === 'object') {
+//         return o;
+//       }
+//     } catch (e) {
+//       console.log({ error: e });
+//     }
+//     return false;
+//   }
