+++ app/modals/ServicesModal.tsx
@@ -24,9 +24,6 @@ const ServicesModal: React.FC<ServicesModalProps> = React.memo(({ i, app }) => {
   console.log('ServicesModal current props (index, app): ', i, ' ', app);
 
   const { user, applications } = useContext(DashboardContext);
-  console.log('user from Dashboard Context:', user);
-  console.log('applications from Dashboard Context: ', applications);
-  console.log('applications[i][2]: ', applications[i][2]);
   const { servicesData, connectToDB } = useContext(ApplicationContext);
   const [services, setServices] = useState<Array<string>>([]);
 
@@ -45,8 +42,6 @@ const ServicesModal: React.FC<ServicesModalProps> = React.memo(({ i, app }) => {
   // adding database type to make connection and fetchServiceNames more efficient
   useEffect(() => {
     console.log('Hi, inside useEffect in ServicesModal. Calling connectToDB function.');
-    console.log("Passing the following parameters for user, i, app, applications, ");
-    console.log(user, ' ', i, ' ', app, ' ', applications, applications[i][1]);
     connectToDB(user, i, app, applications[i][2], applications[i][1]);
   }, [i]);
