+++ electron/routes/data.ts
@@ -3,43 +3,31 @@
 import { ipcMain } from 'electron';
 import fs from 'fs';
 import path from 'path';
+const log = require('electron-log');
+const mongoose = require('mongoose');
+const AWS = require('aws-sdk');
+import fetch, { FetchError } from 'electron-fetch';
+import { fetchData } from './dataHelpers';
 import connectPostgres from '../databases/postgres';
 import connectMongo from '../databases/mongo';
+// import mongodb models:
 import CommunicationModel from '../models/CommunicationsModel';
 import HealthModelFunc from '../models/HealthModel';
 import ServicesModel from '../models/ServicesModel';
 import DockerModelFunc from '../models/DockerModel';
 import KafkaModel from '../models/KafkaModel';
-import fetch, { FetchError } from 'electron-fetch';
-//import { postgresFetch, mongoFetch }  from './dataHelpers';
-import { fetchData } from './dataHelpers';
 import MetricsModel from '../models/MetricsModel';
-const log = require('electron-log');
-
-
-const mongoose = require('mongoose');
-const User = require('../models/UserModel')
-
-// const testURL = 'mongodb+srv://seconddbtest:seconddbtest@cluster0.yhztme0.mongodb.net/?retryWrites=true&w=majority';
-// const connectMongoose = async (i: number, URI: string) => {
-//   try {
-//     const db2 = mongoose.createConnection(testURL);
-//     console.log('connection to user provided db established..');
-//     return db2;
-//   } catch (error) {
-//     console.log('Error connecting to second db... ', error);
-//   }
-// }
-
+const User = require('../models/UserModel');
 
 
 const mongoFetch = fetchData.mongoFetch;
 const postgresFetch = fetchData.postgresFetch;
-const AWS = require('aws-sdk');
+
 
 require('dotenv').config({
    path: path.join(__dirname, './.env'),
 });
+
 // Initiate pool variable for SQL setup
 let pool: any;
 
@@ -49,13 +37,15 @@ let currentDatabaseType: string;
 // Provide location to settings.json
 const settingsLocation = path.resolve(__dirname, '../../settings.json');
 
-// v10 notes: should only be calling connect for local instances. Currently, the services array is differenct for
-// local instances vs cloud instances but this function can still be called by cloud isntances, causing an issue.
-// fix below is a band-aid for now, a better solution would be optimal.
+// v10 notes: should only be calling connect for local instances. Currently, the services array is different for
+// local instances vs cloud instances but this function can still be called by cloud instances, causing an issue.
+// fix below is a band-aid for now, a better solution would be optimal. Multiple calls to different SQL databases has not been tested.
 /**
- * @event   connect
+ * @event   connect - being called by ipcRenderer in ApplicationContext.tsx to establish connection to user provided database
+ *          for local instances.
  * @desc    Connects user to database and sets global currentDatabaseType which
- *          is accessed in info.commsData and info.healthData
+ *          is referenced throughout ipcMain.on functions
+ * @returns (to ipcMain) true or false
  */
 
 ipcMain.on('connect', async (message: Electron.IpcMainEvent, username: string, index: number, URI: string, databaseType: string) => {
@@ -97,10 +87,12 @@ ipcMain.on('connect', async (message: Electron.IpcMainEvent, username: string, i
   }
 });
 
-
+// v10 notes: could improve error handling
 /**
- * @event   serviceRequest/serviceResponse
+ * @event   serviceRequest/serviceResponse - called by ipcRenderer.on in ApplicationContext.tsx to set array of services
+ *          being monitored for local instances. 
  * @desc    Query to services table for all microservices of a specific app
+ * @returns (to ipcMain) an array of queried results of user services
  */
 ipcMain.on('servicesRequest', async (message: Electron.IpcMainEvent) => {
   try {
@@ -122,7 +114,6 @@ ipcMain.on('servicesRequest', async (message: Electron.IpcMainEvent) => {
       result = result.rows;
     }
 
-    // console.log('Sending servicesResponse to frontend with the following result:', result);
     // Async event emitter - send response
     message.sender.send('servicesResponse', JSON.stringify(result));
     // eslint-disable-next-line no-shadow
@@ -131,9 +122,11 @@ ipcMain.on('servicesRequest', async (message: Electron.IpcMainEvent) => {
   }
 });
 
+// v10 notes: SQL instances not tested.
 /**
- * @event   commsRequest/commsResponse
- * @desc    Query for all communication data
+ * @event   commsRequest/commsResponse - called by CommsContext.tsx
+ * @desc    Queries communication data for local instances
+ * @returns queried result 
  */
 ipcMain.on('commsRequest', async (message: Electron.IpcMainEvent) => {
   try {
@@ -164,6 +157,7 @@ ipcMain.on('commsRequest', async (message: Electron.IpcMainEvent) => {
 /**
  * @event   healthRequest/healthResponse
  * @desc    Query for health data for a particular microservice (last 50 data points)
+ * @returns 'healthResponse' and queried results
  */
 ipcMain.on('healthRequest', async (message: Electron.IpcMainEvent, service: string) => {
   try {
@@ -191,8 +185,9 @@ ipcMain.on('healthRequest', async (message: Electron.IpcMainEvent, service: stri
 });
 
 /**
- * @event   dockerRequest/DockerResponse
+ * @event   dockerRequest/dockerResponse
  * @desc    Query for health data for a particular microservice (last 50 data points)
+ * @returns 'dockerResponse' and queried results
  */
 ipcMain.on('dockerRequest', async (message, service) => {
   try {
@@ -230,6 +225,11 @@ ipcMain.on('dockerRequest', async (message, service) => {
   }
 });
 
+/**
+ * @event   savedMetricsRequest/savedMetricsResponse
+ * @desc    Query for metrics saved by user 
+ * @returns 'savedMetricsRequest' and queried results
+ */
 // This event allows the user to change which metrics are saved by the database, so that their database doesn't get bloated with unnecessary data that they don't actually want.
 ipcMain.on('savedMetricsRequest', async (message: Electron.IpcMainEvent) => {
   try {
@@ -256,6 +256,11 @@ ipcMain.on('savedMetricsRequest', async (message: Electron.IpcMainEvent) => {
   }
 });
 
+/**
+ * @event   updateSavedMetrics
+ * @desc    updates saved metrics in user databased
+ * @returns 'dockerResponse' and queried results
+ */
 ipcMain.on('updateSavedMetrics', async (message: Electron.IpcMainEvent, args: Object[]) => {
   try {
     // Mongo Database
@@ -283,33 +288,34 @@ ipcMain.on('updateSavedMetrics', async (message: Electron.IpcMainEvent, args: Ob
   }
 });
 
+
+// // start fetch
+// function extractWord(str: string) {
+//   const res: any[] = [];
+//   const arr = str.split('\n'); // `/\n/`
+//   for (const element of arr) {
+//     if (
+//       element &&
+//       element.length !== 0 &&
+//       element[0] !== '#' &&
+//       element.substring(0, 3) !== 'jmx' &&
+//       element.substring(0, 4) !== "'jmx"
+//     ) {
+//       const metric = element.split(' ')[0];
+//       const metricValue = Number(element.split(' ')[1]);
+//       const time = Date.now();
+//       const temp = { metric: metric, category: 'Event', value: metricValue, time: time };
+//       res.push(temp);
+//     }
+//   }
+//   return res;
+// }
+
 /**
- * @event   eventRequest/EventResponse
- * @desc
+ * @event   kafkaRequest/kafkaResponse
+ * @desc    makes call using mongoFetch or postgresFetch which returns kafka queried results
+ * @returns 'kafkaResponse' and queried results
  */
-
-// start fetch
-function extractWord(str: string) {
-  const res: any[] = [];
-  const arr = str.split('\n'); // `/\n/`
-  for (const element of arr) {
-    if (
-      element &&
-      element.length !== 0 &&
-      element[0] !== '#' &&
-      element.substring(0, 3) !== 'jmx' &&
-      element.substring(0, 4) !== "'jmx"
-    ) {
-      const metric = element.split(' ')[0];
-      const metricValue = Number(element.split(' ')[1]);
-      const time = Date.now();
-      const temp = { metric: metric, category: 'Event', value: metricValue, time: time };
-      res.push(temp);
-    }
-  }
-  return res;
-}
-
 ipcMain.on('kafkaRequest', async message => {
   try {
     let result: any;
@@ -331,7 +337,11 @@ ipcMain.on('kafkaRequest', async message => {
   }
 });
 
-// JJ-ADDITION
+/**
+ * @event   kubernetesRequest/kafkaResponse
+ * @desc    makes call using mongoFetch or postgresFetch
+ * @returns 'kubernetesResponse' and queried results
+ */
 ipcMain.on('kubernetesRequest', async message => {
   try {
     let result: any;
@@ -352,249 +362,15 @@ ipcMain.on('kubernetesRequest', async message => {
   }
 });
 
-/**
- * @event   event ec2MetricsRequest
- * @desc    Connects user to Cloudwatch using aws-sdk and fetches data for EC2 instances
- *          with the passed in parameters
- * @params  username: current user from DashboardContext
- *          appIndex: clicked card's index in the application array from ApplicationContext
+
+  /**
+ * @event   awsAppInfoRequest - invoked in fetchAwsAppInfo in ipcRenderer
+ * @desc    Connects to user or guest database and returns a reponse object with the typeOfService,
+ *          region, and awsURL of the services at provided appIndex.
+ * @params  username: 
+ *          index: 
+ * @returns 'awsAppInfoResponse' to ipcRenderer and a response object containing the typeOfService, region, and awsURL for a selected AWS App
  */
-ipcMain.on(
-  'ec2MetricsRequest',
-  async (message: Electron.IpcMainEvent, username: string, appIndex: number) => {
-    try {
-      // find the clicked service in settings
-      const fileContents = JSON.parse(fs.readFileSync(settingsLocation, 'utf8'));
-      const userAwsService = fileContents[username]?.services[appIndex];
-
-      // grab variables needed in the parameters for Cloudwatch
-      const [region, typeOfService, instanceId, accessKey, secretAccessKey] = [
-        userAwsService[2],
-        userAwsService[4],
-        userAwsService[6],
-        userAwsService[7],
-        userAwsService[8],
-      ];
-
-      // create a new Cloudwatch instance
-      const cloudwatch = new AWS.CloudWatch({
-        region: region,
-        accessKeyId: accessKey,
-        secretAccessKey: secretAccessKey,
-      });
-
-      // metrics being requested. Each metric will get its own graph in AwsGraphContainer. List of metrics available can be found in AWS Documentation
-      const metricsNamesArray = ['CPUUtilization', 'NetworkIn', 'NetworkOut', 'DiskReadBytes'];
-
-      const paramsArray = metricsNamesArray.map(metric => {
-        // param needed for Cloudwatch to fetch data, highly opinionated boilerplate
-        const params = {
-          EndTime: new Date(),
-          MetricName: metric,
-          Namespace: typeOfService,
-          Period: 60,
-          StartTime: new Date(new Date().getTime() - 60 * 60 * 1000),
-          Statistics: ['Average'],
-          Dimensions: [
-            {
-              Name: 'InstanceId',
-              Value: instanceId,
-            },
-          ],
-        };
-
-        return params;
-      });
-
-      const fetchData = async () => {
-        const fetched = {};
-
-        for (let i = 0; i < paramsArray.length; i++) {
-          // getMetricsStatistics: the fetch method using the current params
-          const data = await cloudwatch.getMetricStatistics(paramsArray[i]).promise();
-
-          // transform data in format for frontend to render graphs
-          const newData = data.Datapoints.map((el, index: number) => {
-            let transformedData = {};
-
-            (transformedData['time'] = data.Datapoints[index].Timestamp),
-              (transformedData['metric'] = data.Label),
-              (transformedData['value'] = data.Datapoints[index].Average),
-              (transformedData['unit'] = data.Datapoints[index].Unit);
-
-            return transformedData;
-          });
-
-          fetched[paramsArray[i].MetricName] = newData;
-        }
-
-        return fetched;
-      };
-
-      fetchData().then(data => {
-        message.sender.send('ec2MetricsResponse', JSON.stringify(data)); // send data to frontend
-      });
-    } catch (err) {
-      console.log('Error in "ec2MetricsRequest" event', message);
-      message.sender.send('ec2MetricsResponse', {
-        CPUUtilization: [],
-        NetworkIn: [],
-        NetworkOut: [],
-        DiskReadBytes: [],
-      });
-    }
-  }
-);
-
-/**
- * @event   event ecsMetricsRequest
- * @desc    Connects user to Cloudwatch using aws-sdk and fetches data for ECS Clusters/Services
- *          with the passed in parameters
- * @params  username: current user from DashboardContext
- *          appIndex: clicked card's index in the application array from ApplicationContext
- */
-ipcMain.on(
-  'ecsMetricsRequest',
-  async (message: Electron.IpcMainEvent, username: string, appIndex: number) => {
-    try {
-      // similar function architecture as EC2 fetch request
-      const fileContents = JSON.parse(fs.readFileSync(settingsLocation, 'utf8'));
-      const userAwsService = fileContents[username]?.services[appIndex];
-
-      const [region, typeOfService, accessKey, secretAccessKey] = [
-        userAwsService[2],
-        userAwsService[4],
-        userAwsService[7],
-        userAwsService[8],
-      ];
-
-      const cloudwatch = new AWS.CloudWatch({
-        region: region,
-        accessKeyId: accessKey,
-        secretAccessKey: secretAccessKey,
-      });
-
-      const listMetricsParams = {
-        Namespace: 'AWS/ECS',
-      };
-
-      let clusterName: string = '';
-      const serviceNames: string[] = [];
-      const ecsData = {};
-
-      cloudwatch.listMetrics(listMetricsParams, (err, data) => {
-        if (err) {
-          console.log('Error', err);
-        } else {
-          for (let i = 0; i < data.Metrics.length; i++) {
-            const dimensions: any[] = data.Metrics[i].Dimensions;
-
-            for (let j = 0; j < dimensions.length; j++) {
-              if (dimensions[j].Name === 'ClusterName') {
-                clusterName = dimensions[j].Value;
-              }
-
-              if (dimensions[j].Name === 'ServiceName') {
-                if (!serviceNames.includes(dimensions[j].Value)) {
-                  serviceNames.push(dimensions[j].Value);
-                }
-              }
-            }
-          }
-        }
-
-        const fetchData = async () => {
-          if (clusterName !== '' && serviceNames.length !== 0) {
-            for (let i = 0; i < serviceNames.length; i++) {
-              let currentService = serviceNames[i];
-
-              // cluster may have multiple services, and each service needs its own parameter
-              const params = {
-                MetricDataQueries: [
-                  {
-                    Id: 'm1',
-                    MetricStat: {
-                      Metric: {
-                        Namespace: 'AWS/ECS',
-                        MetricName: 'CPUUtilization',
-                        Dimensions: [
-                          {
-                            Name: 'ClusterName',
-                            Value: clusterName,
-                          },
-                          {
-                            Name: 'ServiceName',
-                            Value: currentService,
-                          },
-                        ],
-                      },
-                      Period: 60,
-                      Stat: 'Average',
-                    },
-                    ReturnData: true,
-                  },
-                  {
-                    Id: 'm2',
-                    MetricStat: {
-                      Metric: {
-                        Namespace: 'AWS/ECS',
-                        MetricName: 'MemoryUtilization',
-                        Dimensions: [
-                          {
-                            Name: 'ClusterName',
-                            Value: clusterName,
-                          },
-                          {
-                            Name: 'ServiceName',
-                            Value: currentService,
-                          },
-                        ],
-                      },
-                      Period: 60,
-                      Stat: 'Average',
-                    },
-                    ReturnData: true,
-                  },
-                ],
-                StartTime: new Date(Date.now() - 60 * 60 * 1000),
-                EndTime: new Date(),
-                ScanBy: 'TimestampDescending',
-              };
-
-              const data = await cloudwatch.getMetricData(params).promise();
-
-              ecsData[currentService] = {
-                CPUUtilization: {
-                  value: data.MetricDataResults[0].Values,
-                  time: data.MetricDataResults[0].Timestamps,
-                },
-                MemoryUtilization: {
-                  value: data.MetricDataResults[1].Values,
-                  time: data.MetricDataResults[1].Timestamps,
-                },
-              };
-            }
-          }
-
-          ecsData['clusterInfo'] = {
-            clusterName: clusterName,
-            region: region,
-            typeOfService: typeOfService,
-          };
-          return ecsData;
-        };
-
-        fetchData().then(data => {
-          message.sender.send('ecsMetricsResponse', JSON.stringify(data));
-        });
-      });
-    } catch (err) {
-      console.log(err);
-    }
-  }
-);
-
-// returns response object containing the typeOfService, region, and awsURL for a selected AWS APP
 ipcMain.on(
   'awsAppInfoRequest',
   async (message: Electron.IpcMainEvent, username: string, appIndex: number) => {
@@ -637,86 +413,55 @@ ipcMain.on(
   }
 );
 
-ipcMain.on('eksMetricsRequest', async (message:Electron.IpcMainEvent, username: string, appIndex: number) => {
-  const fileContents = JSON.parse(fs.readFileSync(settingsLocation, 'utf8'));
-  const userAwsService = fileContents[username]?.services[appIndex];
-  const [typeOfService, region, awsUrl] = [userAwsService[4], userAwsService[2], userAwsService[9]];
-  const url = `${awsUrl}/api/search?folderIds=0`
-  console.log("hi")
-  const response = await fetch(url, {
-    method: 'GET',
-    headers: {
-      "Access-Control-Allow-Origin": "*",
-      Accept: "application/json",
-      "Content-Type": "application/json",
-      Authorization: 'Bearer eyJrIjoiamN4UGRKVHg3cUQyZ201N042NW41bHg5bGhJaVFlaFciLCJuIjoidGVzdEtleSIsImlkIjoxfQ=='
-    },
-  });
-  log.info(response);
-  const data = await response.json();
-  console.log(data)
-  console.log(awsUrl)
-  message.sender.send('eksMetricsResponse', JSON.stringify(data));
-});
+// v10 notes: function that was left over in data.ts. Have not found any instance where it is
+// invoked but did not want to delete unless it is needed by another team. 
+// // start fetch
+// function extractWord(str: string) {
+//   const res: any[] = [];
+//   const arr = str.split('\n'); // `/\n/`
+//   for (const element of arr) {
+//     if (
+//       element &&
+//       element.length !== 0 &&
+//       element[0] !== '#' &&
+//       element.substring(0, 3) !== 'jmx' &&
+//       element.substring(0, 4) !== "'jmx"
+//     ) {
+//       const metric = element.split(' ')[0];
+//       const metricValue = Number(element.split(' ')[1]);
+//       const time = Date.now();
+//       const temp = { metric: metric, category: 'Event', value: metricValue, time: time };
+//       res.push(temp);
+//     }
+//   }
+//   return res;
+// }
 
-//end fetch
+//   /**
+//  * @event   awsAppInfoRequest - invoked in fetchAwsAppInfo in ipcRenderer
+//  * @desc    Connects to user or guest database and returns a reponse object with the typeOfService,
+//  *          region, and awsURL of the services at provided appIndex.
+//  * @params  username: 
+//  *          index: 
+//  */
 // ipcMain.on(
-//   'eksMetricsRequest',
+//   'awsAppInfoRequest',
 //   async (message: Electron.IpcMainEvent, username: string, appIndex: number) => {
 //     try {
-//       // similar function architecture as EC2 fetch request
 //       const fileContents = JSON.parse(fs.readFileSync(settingsLocation, 'utf8'));
 //       const userAwsService = fileContents[username]?.services[appIndex];
-//       const awsUrl = userAwsService[9];
-//       const url = `${awsUrl}/api/search?folderIds=0`
 
-
-//         const fetchData = async () => {
-//              await fetch(url, {
-//                 method: 'GET',
-//                 headers: {
-//                   "Access-Control-Allow-Origin": "*",
-//                   Accept: "application/json",
-//                   "Content-Type": "application/json",
-//                   Authorization: 'Bearer eyJrIjoiamN4UGRKVHg3cUQyZ201N042NW41bHg5bGhJaVFlaFciLCJuIjoidGVzdEtleSIsImlkIjoxfQ=='
-//                 },
-//               });
-
-//         fetchData().then(data => {
-//           message.sender.send('eksMetricsResponse', JSON.stringify(data));
-//         });
+//       const [typeOfService, region, awsUrl] = [userAwsService[4], userAwsService[2], userAwsService[9]];
+//       const response = {
+//         typeOfService: typeOfService,
+//         region: region,
+//         awsUrl: awsUrl
 //       };
+
+//       message.sender.send('awsAppInfoResponse', JSON.stringify(response));
 //     } catch (err) {
-//       console.log(err);
+//       console.log('Error in awsAppInfoRequest', message);
+//       message.sender.send('awsAppInfoResponse', { typeOfService: '', region: '' });
 //     }
 //   }
 // );
-
-  /**
- * @event   awsAppInfoRequest - invoked in fetchAwsAppInfo in ipcRenderer
- * @desc    Connects to user or guest database and returns a reponse object with the typeOfService,
- *          region, and awsURL of the services at provided appIndex.
- * @params  username: 
- *          index: 
- */
-ipcMain.on(
-  'awsAppInfoRequest',
-  async (message: Electron.IpcMainEvent, username: string, appIndex: number) => {
-    try {
-      const fileContents = JSON.parse(fs.readFileSync(settingsLocation, 'utf8'));
-      const userAwsService = fileContents[username]?.services[appIndex];
-
-      const [typeOfService, region, awsUrl] = [userAwsService[4], userAwsService[2], userAwsService[9]];
-      const response = {
-        typeOfService: typeOfService,
-        region: region,
-        awsUrl: awsUrl
-      };
-
-      message.sender.send('awsAppInfoResponse', JSON.stringify(response));
-    } catch (err) {
-      console.log('Error in awsAppInfoRequest', message);
-      message.sender.send('awsAppInfoResponse', { typeOfService: '', region: '' });
-    }
-  }
-);
